Share stack pool between threads
